Refactor to share code between EditorSession and ModalEditorSession.
